modified:   main/nofutureball/src/nofutureball/Animatable.java 	modified:   main/nofutureball/src/nofutureball/Container.java 	modified:   main/nofutureball/src/nofutureball/Game.java 	renamed:    main/nofutureball/src/nofutureball/gameObject.java -> main/nofutureball/src/nofutureball/GameObject.java 	modified:   main/nofutureball/src/nofutureball/NoFutureBall.java 	modified:   main/nofutureball/src/nofutureball/Player.java 	modified:   main/nofutureball/src/nofutureball/Room.java 	modified:   main/nofutureball/src/nofutureball/Wall.java

+++ main/nofutureball/src/nofutureball/Animatable.java
@@ -25,8 +25,8 @@ public abstract class Animatable extends Entity {
 		super.render();
 		if (animations != null) {
 			animations.getCurrentlyPlaying().draw(
-					(int) (position.x + offset.x - pivot.x),
-					(int) (position.y + offset.y - pivot.y),
+					 (position.x + offset.x - pivot.x),
+					 (position.y + offset.y - pivot.y),
 					new Color(1f, 1f, 1f));
 		} else {
 			System.out
